delete blog api created

+++ src/controlers/deleteControler/deleteJobApplicationControler.js
@@ -7,25 +7,36 @@ const deleteJobApplicationControler = async (req, res) => {
     const query = { jobId: id };
 
     const isAvailable = await jobApplicationCollections.findOne(query);
+
     if (!isAvailable) {
       return res.send({
         success: false,
         message: "Your already canceled your application.",
       });
-    } else {
-      const result = await jobApplicationCollections.deleteOne(query);
-      if (!result.acknowledged) {
-        return res.send({
-          success: false,
-          message: "somethig went wrong you cant delete you job post yet.",
-        });
-      } else {
-        res.send({
-          success: true,
-          message: "You successfully canceled your job application.",
-        });
-      }
     }
+
+    // verify jwt with email address
+    if (isAvailable.userEmail !== req.decoded) {
+      return res.send({
+        success: false,
+        message:
+          "Unauthorized access.You are not a valid user for delete this job application.",
+      });
+    }
+
+    const result = await jobApplicationCollections.deleteOne(query);
+
+    if (!result.acknowledged) {
+      return res.send({
+        success: false,
+        message: "somethig went wrong you cant delete you job post yet.",
+      });
+    }
+
+    res.send({
+      success: true,
+      message: "You successfully canceled your job application.",
+    });
   } catch (error) {
     res.send({
       success: false,
